\name{armatest}
\alias{armatest}

\title{Test RcppArmadillo exp}

\description{

Test only

}

\usage{

armatest(N = 200, K = 100, ncores = 1, debug = FALSE)

}

\arguments{

  \item{N}{nrows}
  \item{K}{ncols}
  \item{ncores}{integer cluster size}
  \item{debug}{logical}

}

\details{

None yet

}

\value{

Logical (TRUE if all exp positive)

}

